style: 반응형 구현 (모바일 화면 제외)
