+++ apps/client/src/app/(member)/login/page.tsx
@@ -1,5 +1,5 @@
-import Login from '../../../components/form/Login';
 import AuthFrame from '../../../components/pages/member/AuthFrame';
+import Login from '../../../components/pages/member/Login';
 
 export default function page() {
   return (
