--- apps/client/src/components/form/index.css
@@ -1,11 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-.scrollbar-hide {
-  scrollbar-width: none; /* Firefox */
-}
-
-.scrollbar-hide::-webkit-scrollbar {
-  display: none; /* Chrome, Safari */
-}
